requirements.txt
django==2.2.19
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
drf-extra-fields==3.2.1
django-filter==21.1
drf-extra-fields-3.4.1
